+++ pkg/cmd/console/clusterclaim/cmd.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 
 	genericclioptionscm "github.com/open-cluster-management/cm-cli/pkg/genericclioptions"
+	"k8s.io/kubectl/pkg/cmd/get"
 	cmdutil "k8s.io/kubectl/pkg/cmd/util"
 
 	"github.com/open-cluster-management/cm-cli/pkg/helpers"
@@ -45,8 +46,13 @@ func NewCmd(cmFlags *genericclioptionscm.CMFlags, streams genericclioptions.IOSt
 		// },
 	}
 
+	o.GetOptions.PrintFlags = get.NewGetPrintFlags()
+
+	o.GetOptions.PrintFlags.AddFlags(cmd)
+
 	cmd.Flags().StringVar(&o.ClusterPoolHost, "cph", "", "The clusterpoolhost to use")
 	cmd.Flags().IntVar(&o.Timeout, "timeout", 60, "Timeout to get the cluster claim running")
+	cmd.Flags().BoolVar(&o.WithCredentials, "creds", o.WithCredentials, "If set the credentials will be displayed")
 
 	return cmd
 }