+++ pkg/cmd/create/clusterclaim/options.go
@@ -4,6 +4,7 @@ package clusterclaim
 import (
 	genericclioptionscm "github.com/open-cluster-management/cm-cli/pkg/genericclioptions"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
+	"k8s.io/kubectl/pkg/cmd/get"
 )
 
 type Options struct {
@@ -13,6 +14,8 @@ type Options struct {
 	ClusterClaims   string
 	ClusterPool     string
 	ClusterPoolHost string
+	GetOptions      *get.GetOptions
+	WithCredentials bool
 	Timeout         int
 	SkipSchedule    bool
 	//The file to output the resources will be sent to the file.
@@ -21,6 +24,7 @@ type Options struct {
 
 func newOptions(cmFlags *genericclioptionscm.CMFlags, streams genericclioptions.IOStreams) *Options {
 	return &Options{
-		CMFlags: cmFlags,
+		CMFlags:    cmFlags,
+		GetOptions: get.NewGetOptions("cm", streams),
 	}
 }