+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 # Release Content
 ## Additions
 - (Add info to the cm attach command #120)[https://github.com/open-cluster-management/cm-cli/issues/120]
+- (Add flag --creds to retrieve credentials while running 'run', 'console', 'create' clusterclain #129)[https://github.com/open-cluster-management/cm-cli/issues/129]
 - Support Multi Cluster Engine
 ## Breaking changes